% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_help.R
\name{pick_fgeo_concept}
\alias{pick_fgeo_concept}
\alias{pick_package}
\alias{pick_concept}
\alias{pick_fgeo_package}
\title{Pick documentation matching some concepts or package names.}
\usage{
pick_package(x, cols = NULL, url = "https://forestgeo.github.io/")

pick_concept(x, cols = NULL, url = "https://forestgeo.github.io/")

pick_fgeo_concept(x)

pick_fgeo_package(x)
}
\arguments{
\item{x}{A character vector giving concepts or package names to match.}

\item{cols}{Character vector giving the name of columns to select from the
result, e.g.  c("alias", "topic", "title").}

\item{url}{Character vector of length-1 giving a base url, e.g.
"https://forestgeo.github.io/".}
}
\value{
A dataframe.
}
\description{
These functions help you to reference the documentation of specific functions
across multiple packages. This is useful in meta-packages.
}
\examples{
pick_fgeo_concept("datasets")
pick_fgeo_package("fgeo.x")

}
