% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_help.R
\name{pick_package}
\alias{pick_package}
\alias{pick_concept}
\alias{reference_package}
\alias{reference_concept}
\title{Pick documentation matching some concepts or package names.}
\usage{
pick_package(x, url = NULL, packages = NULL)

pick_concept(x, url = NULL, packages = NULL)

reference_package(x, url = NULL, packages = NULL,
  strip_s3class = TRUE)

reference_concept(x, url = NULL, packages = NULL,
  strip_s3class = TRUE)
}
\arguments{
\item{x}{A character vector giving concepts or package names to match.}

\item{url}{Character vector of length-1 giving a base url, e.g.
"https://forestgeo.github.io/".}

\item{packages}{A character vector with the names of packages to search
through, or NULL in which case all available packages in the library trees
specified by lib.loc are searched. Defaults to search only the core
packages of \strong{fgeo}.}

\item{strip_s3class}{\code{TRUE} removes the class component of S3 methods.}
}
\value{
A dataframe.
}
\description{
These functions help you to reference the documentation of specific functions
across multiple packages. This is useful in meta-packages. * The functions
ending in \code{package} output all exported objects of the given packages, and
the functions ending in \code{concept} output all objects containing the given
concepts (strings passed to \code{@family}).
\itemize{
\item The functions starting with \code{reference_} output a dataframe optimized for
building a reference table (maybe with \code{DT::datatable()} or
\code{knitr::kable()}).
}
}
\examples{
pick_package("base")

pick_concept("PCA")

packages <- c("stats", "MASS")
pick_package(packages)

concepts <- c("combine strings", "files", "PCA")
pick_concept(concepts)

\dontrun{
if (requireNamespace("fgeo.x", quietly = TRUE)) {
  library(fgeo.x)

  reference_concept("datasets", url = "https://forestgeo.github.io/")
  reference_package("datasets", url = "https://forestgeo.github.io/")
}
}
}
