% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdoc-package.R
\docType{package}
\name{pkgdoc-package}
\alias{pkgdoc}
\alias{pkgdoc-package}
\title{pkgdoc: Dataframe and Reference the Documentation of Available
    Packages}
\description{
Dataframe the documentation of available packages,
    and reference objects across multiple packages by package or concept.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/pkgdoc}
  \item Report bugs at \url{https://github.com/maurolepore/pkgdoc/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

}
\keyword{internal}
