% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_concept.R
\name{reference_concept}
\alias{reference_concept}
\title{Pick documentation matching some concepts.}
\value{
A dataframe.
}
\description{
Create a data.frame with documentation metadata of one or more packages. If
you develop a universe, such as the tidyverse or tidymodels, pkgdoc helps you
to easily create a universe-wide reference for the pkgdown website of your
meta-package.
}
\examples{
reference_concept("Datasets available by data()")
}
\seealso{
Other retrieve documentation: 
\code{\link{search_docs}()}
}
