% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_strip_class.R
\name{s3_strip_class}
\alias{s3_strip_class}
\title{Strip the class part of attached S3 methods (do nothing with other functions).}
\usage{
s3_strip_class(x)
}
\arguments{
\item{x}{A character vector giving candidate S3 methods.}
}
\value{
Character.
}
\description{
Strip the class part of attached S3 methods (do nothing with other functions).
}
\examples{
s3_strip_class(c("print", "print.factor", "print.data.frame"))

}
\seealso{
utils::isS3method()
}
