% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdoc-package.R
\docType{package}
\name{pkgdoc-package}
\alias{pkgdoc}
\alias{pkgdoc-package}
\title{pkgdoc: Reference Documented R Objects Across Multiple Packages}
\description{
Reference documented R objects across multiple
    packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/pkgdoc}
  \item Report bugs at \url{https://github.com/maurolepore/pkgdoc/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

}
\keyword{internal}
