% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_package.R
\name{reference_package}
\alias{reference_package}
\title{Create a data.frame with documentation metadata of one or more packages}
\usage{
reference_package(
  x,
  url_template = NULL,
  url = NULL,
  packages = NULL,
  strip_s3class = TRUE
)
}
\arguments{
\item{x}{A character vector giving concepts or package names to match.}

\item{url_template}{Character. A template to generate links to documentation
based on the column names of the output -- typically \code{package} and \code{topic},
e.g. \code{"https://maurolepore.github.io/{package}/reference/{topic}.html"}
(\code{glue::glue()} syntax).}

\item{url}{Character vector of length-1 giving a base url, e.g.
"https://maurolepore.github.io/".}

\item{packages}{A character vector with the names of packages to limit the
search, or \code{NULL} to serach in all known libraries.}

\item{strip_s3class}{\code{TRUE} removes the class component of S3 methods.}
}
\value{
A dataframe.
}
\description{
Create a data.frame with documentation metadata of one or more packages. If
you develop a universe, such as the tidyverse or tidymodels, pkgdoc helps you
to easily create a universe-wide reference for the pkgdown website of your
meta-package.
}
\examples{
reference_package("datasets")

}
\seealso{
Other retrieve documentation: 
\code{\link{search_docs}()}
}
\concept{retrieve documentation}
