% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_help.R
\name{search_docs}
\alias{search_docs}
\title{Dataframe the documentation of installed packages.}
\usage{
search_docs(packages = NULL)
}
\arguments{
\item{packages}{A character vector with the names of packages to search
through, or NULL in which case all available packages in the library trees
specified by lib.loc are searched. Defaults to search only the core
packages of \strong{fgeo}.}
}
\value{
A dataframe.
}
\description{
This is fundamentally a dataframe version of the output of
\code{\link[utils:hsearch_db]{utils::hsearch_db()}}.
}
\examples{
search_docs(NULL)

}
